+++ .storybook/config.js
@@ -66,6 +66,7 @@ configure(() => {
   require('./../packages/bpk-component-textarea/stories');
   require('./../packages/bpk-component-ticket/stories');
   require('./../packages/bpk-component-tile/stories');
+  require('./../packages/bpk-component-toast/stories');
   require('./../packages/bpk-component-tooltip/stories');
   require('./../packages/bpk-mixins/stories');
 }, module);